% Generated by roxygen2 (4.0.2): do not edit by hand
\name{activate_core}
\alias{activate_core}
\title{Activate multiple CPU cores for parallel computing.}
\usage{
activate_core(n_core = 4, verbose = TRUE)
}
\arguments{
\item{n_core}{Number of CPU cores to be activated for parallel computing.}

\item{verbose}{Display messages (TRUE or FALSE)}
}
\description{
Activate multiple CPU cores for parallel computing.
}
\examples{
activate_core(4)
}

