% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o_feaSelect}
\alias{h2o_feaSelect}
\title{Feature Selection using H2O Random Forest}
\usage{
h2o_feaSelect(x_train, y_train, type = "classification", n_fold = 5,
  n_repeat = 3, n_tree = 100, n_size = c(10:15), n_seed = 1234,
  n_threads = -1, plot_graph = TRUE, balance = TRUE)
}
\arguments{
\item{x_train}{predictors}

\item{y_train}{labels}
}
\description{
Feature Selection using H2O Random Forest
}
\examples{
## Select Best Features
h2o_feaSelect(x_train, y_train)
}

