% Generated by roxygen2 (4.0.2): do not edit by hand
\name{install_h2o}
\alias{install_h2o}
\title{Install/Upgrade H2O Package and JVM to the latest or a specific bleeding edge version}
\usage{
install_h2o(h2o_ver = NULL, force = FALSE)
}
\arguments{
\item{h2o_ver}{Specific a version number. If left empty (null), it will install the latest version.}
}
\description{
Install/Upgrade H2O Package and JVM to the latest or a specific bleeding edge version
}
\examples{
## Install/upgrade h2o to the latest bleeding edge version
install_h2o()

## Install a specific version (e.g. 1500)
install_h2o(h2o_ver = '1500')

## Force overwritting existing version
install_h2o(h2o_ver = '1500', force = TRUE)
}

