% Generated by roxygen2 (4.0.1): do not edit by hand
\name{train_rbm}
\alias{train_rbm}
\title{Extracts hidden features from multiple predictors (x) using RBM and returns the new predictors (x_new) & final RBM (model).}
\usage{
train_rbm(x, n_features = 100, n_batchsize = 100, n_epochs = 100)
}
\arguments{
\item{x}{Multi-column predictors (numeric - matrix).}

\item{n_features}{Number of hidden features to be extracted. It can be single layer (e.g. 100) or multiple layers (e.g. c(400, 400, 20)). (Default: 100)}

\item{n_batchsize}{Size of mini batches for RBM training. (Default: 100)}

\item{n_epochs}{Number of iterative steps required for each RBM training. (Default: 100)}
}
\description{
Extracts hidden features from multiple predictors (x) using RBM and returns the new predictors (x_new) & final RBM (model).
}
\examples{
## Train a single layer of RBM with 100 hidden features
model_rbm <- train_rbm(x, n_features = 100)
x_new <- model_rbm$x_new

## Train three stacked layers of RBM with 400, 400 and 100 hidden features (note: takes a long time)
model_rbm <- train_rbm(x, n_features = c(400, 400, 100))
}

