% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transform_x}
\alias{transform_x}
\title{Transforms original predictors (x) into new predictors (x_new) using a trained RBM model object.}
\usage{
transform_x(model_rbm, x)
}
\arguments{
\item{model_rbm}{A trained RBM object from train_rbm(...).}

\item{x}{Original predictors.}
}
\description{
Transforms original predictors (x) into new predictors (x_new) using a trained RBM model object.
}
\examples{
## Extracts hidden features and create new predictors.
model_rbm <- train_rbm(x, n_features = 10)
x_new <- transform_x(model_rbm, x)
}

