% Generated by roxygen2 (4.0.1): do not edit by hand
\name{upgrade_h2o}
\alias{upgrade_h2o}
\title{Upgrade H2O Package and JVM to the latest or a specific version}
\usage{
upgrade_h2o(h2o_ver = NULL)
}
\arguments{
\item{model_rbm}{A trained RBM object from train_rbm(...).}
}
\description{
Upgrade H2O Package and JVM to the latest or a specific version
}
\examples{
## Upgrade it to the latest version
upgrade_h2o()

## Upgrade it to a specific version
upgrade_h2o(h2o_ver = '1442')
}

