% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stop_timer}
\alias{stop_timer}
\title{Stop the timer and return the time difference in seconds.}
\usage{
stop_timer(time_start)
}
\description{
Stop the timer and return the time difference in seconds.
}
\examples{
tt <- start_timer()
stop_timer(tt)
}

