% Generated by roxygen2 (4.0.2): do not edit by hand
\name{train_rbm}
\alias{train_rbm}
\title{Extracts hidden features from multiple predictors (x) using RBM and returns the final RBM model.}
\usage{
train_rbm(x, n_features = 100, n_batchsize = 100, n_epochs = 100,
  rmv_nearZeroVar = TRUE)
}
\arguments{
\item{x}{Multi-column predictors (numeric - matrix).}

\item{n_features}{Number of hidden features to be extracted. (Default: 100)}

\item{n_batchsize}{Size of mini batches for RBM training. (Default: 100)}

\item{n_epochs}{Number of iterative steps required for each RBM training. (Default: 100)}

\item{rmv_nearZeroVar}{Should the variables of near zero variance to be removed? (Default: TRUE)}
}
\value{
rbm The main RBM object with the weights of hidden features and related stats.

nzv The columns that have near zero variance.

pp The caret object that contains the pre-processing stats.
}
\description{
Extracts hidden features from multiple predictors (x) using RBM and returns the final RBM model.
}
\examples{
## Train a single layer of RBM with 100 hidden features
model_rbm <- train_rbm(x, n_features = 100)

## Transform the original predictors (x) into new predictors (x_new)
x_new <- transform_x(model_rbm, x)
}

